fix: move getAuthedImageUrl location
affects: @esri/hub-common
